Remove separate upper/lower term arguments for faceting
In fact, for most purposes for which I'm aware of "prefix"
feature being used, "target" would actually be a full (though
not drop-in) replacement for the desired functionality.
"target" might even make "prefix" obsolete, but there may
be other use cases out there that I haven't thought of; in
any case, we're not going to extend functionality that we
have no immediate use for, when simply preserving existing
functionality and backward compatibility will suffice.